backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.29
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
